+++ bugsnag-android-core/src/test/java/com/bugsnag/android/BugsnagTestUtils.java
@@ -2,7 +2,7 @@ package com.bugsnag.android;
 
 import org.jetbrains.annotations.NotNull;
 
-import java.util.Collections;
+import java.io.File;
 import java.util.Date;
 import java.util.HashMap;
 
@@ -26,6 +26,22 @@ final class BugsnagTestUtils {
         return convert(generateConfiguration());
     }
 
+    static EventPayload generateEventPayload(ImmutableConfig config) {
+        return new EventPayload(config.getApiKey(), generateEvent(), new Notifier());
+    }
+
+    static Event generateEvent() {
+        Throwable exc = new RuntimeException();
+        Event event = new Event(
+                exc,
+                BugsnagTestUtils.generateImmutableConfig(),
+                HandledState.newInstance(HandledState.REASON_HANDLED_EXCEPTION),
+                NoopLogger.INSTANCE
+        );
+        event.setApp(generateAppWithState());
+        event.setDevice(generateDeviceWithState());
+        return event;
+    }
 
     static ImmutableConfig convert(Configuration config) {
         return ImmutableConfigKt.convertToImmutableConfig(config, null);