+++ bugsnag-android-core/src/test/java/com/bugsnag/android/NativeStackframeSerializationTest.kt
@@ -12,13 +12,16 @@ internal class NativeStackframeSerializationTest {
     companion object {
         @JvmStatic
         @Parameters
-        fun testCases() =
-            generateSerializationTestCases(
+        fun testCases(): Collection<Pair<NativeStackframe, String>> {
+            val last = NativeStackframe(null, null, null, null, null, null)
+            last.type = ErrorType.ANDROID
+            return generateSerializationTestCases(
                 "native_stackframe",
                 NativeStackframe("aMethod", "aFile", 1, 2, 3, 4),
                 NativeStackframe("aMethod", "aFile", 1, null, null, null),
-                NativeStackframe(null, null, null, null, null, null)
+                last
             )
+        }
     }
 
     @Parameter