+++ CHANGELOG.md
@@ -4,7 +4,10 @@
 
 ### Enhancements
 
- * Support native stack traces in the ANR plugin
+* Support setting error type on individual stack frames
+   [#1001](https://github.com/bugsnag/bugsnag-android/pull/1001)
+
+* Support native stack traces in the ANR plugin
    [#972](https://github.com/bugsnag/bugsnag-android/pull/972)
 
 ## 5.2.3 (2020-11-04)