+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## TBD
+## 5.3.0 (2020-12-02)
 
 * Add integrity header to verify Error and Session API payloads have not changed
   [#978](https://github.com/bugsnag/bugsnag-android/pull/978)